+++ bidict/_common.py
@@ -42,12 +42,8 @@ class BidirectionalMapping(Mapping):
 
         Causes conforming classes to be virtual subclasses automatically.
         """
-        if cls is BidirectionalMapping:
-            try:
-                mro = C.__mro__
-            except AttributeError:
-                # Old-style classes in Python 2.7 do not have the __mro__ attribute!
-                return False
+        mro = getattr(C, '__mro__', None)
+        if mro and cls is BidirectionalMapping:
             return all(any(B.__dict__.get(i) for B in mro) for i in cls._subclsattrs)
         return NotImplemented